Move BlockDag into casper/src
