+++ casper/src/main/scala/coop/rchain/casper/BlockDag.scala
@@ -1,0 +1,13 @@
+package coop.rchain.casper
+
+import com.google.protobuf.ByteString
+import coop.rchain.casper.Estimator.{BlockHash, Validator}
+import coop.rchain.casper.protocol.BlockMessage
+
+import scala.collection.immutable.{HashMap, HashSet}
+
+final case class BlockDag(idToBlocks: HashMap[Int, BlockMessage],
+                          blockLookup: HashMap[ByteString, BlockMessage],
+                          childMap: HashMap[BlockHash, HashSet[BlockHash]],
+                          latestMessages: HashMap[Validator, BlockHash],
+                          currentId: Int)
