+++ casper/src/test/scala/coop/rchain/casper/ForkchoiceTest.scala
@@ -1,7 +1,7 @@
 package coop.rchain.casper
 
 import com.google.protobuf.ByteString
-import coop.rchain.casper.internals._
+import coop.rchain.casper.BlockDagState._
 import coop.rchain.casper.protocol.{BlockMessage, Bond}
 import org.scalatest.{FlatSpec, Matchers}
 import coop.rchain.catscontrib._
@@ -19,6 +19,14 @@ import scala.collection.immutable.{HashMap, HashSet}
 
 class ForkchoiceTest extends FlatSpec with Matchers with BlockGenerator {
   type StateWithChain[A] = State[BlockDag, A]
+  val initState =
+    BlockDag(
+      HashMap.empty[Int, BlockMessage],
+      HashMap.empty[BlockHash, BlockMessage],
+      HashMap.empty[BlockHash, HashSet[BlockHash]],
+      HashMap.empty[Validator, BlockHash],
+      0
+    )
 
   // See https://docs.google.com/presentation/d/1znz01SF1ljriPzbMoFV0J127ryPglUYLFyhvsb-ftQk/edit?usp=sharing slide 29 for diagram
   "Estimator on Simple DAG" should "return the appropriate score map and forkchoice" in {
@@ -60,11 +68,6 @@ class ForkchoiceTest extends FlatSpec with Matchers with BlockGenerator {
                                           HashMap(v1 -> b7.blockHash, v2 -> b4.blockHash))
       } yield b8
 
-    val initState =
-      BlockDag(HashMap.empty[Int, BlockMessage],
-            HashMap.empty[BlockHash, BlockMessage],
-            HashMap.empty[BlockHash, HashSet[BlockHash]],
-            0)
     val chain: BlockDag = createChain.runS(initState).value
 
     val genesis = chain.idToBlocks(1)
@@ -127,11 +130,6 @@ class ForkchoiceTest extends FlatSpec with Matchers with BlockGenerator {
                HashMap(v1 -> b6.blockHash, v2 -> b5.blockHash, v3 -> b4.blockHash))
       } yield b8
 
-    val initState =
-      BlockDag(HashMap.empty[Int, BlockMessage],
-            HashMap.empty[BlockHash, BlockMessage],
-            HashMap.empty[BlockHash, HashSet[BlockHash]],
-            0)
     val chain: BlockDag = createChain.runS(initState).value
 
     val genesis = chain.idToBlocks(1)
